test: added conftest files with shared fixtures
