+++ tests/test_core/conftest.py
@@ -1,0 +1,38 @@
+import pytest
+import numpy as np
+
+
+# Create example data with fixtures to test all the functions
+@pytest.fixture()
+def example_xy_data_pos_int() -> np.ndarray:
+    positions = np.array([[1, 2, 3, 4], [5, 6, 7, 8]])
+    return positions
+
+
+@pytest.fixture()
+def example_xy_data_none() -> np.ndarray:
+    positions = np.array([[None, None, None, None], [None, None, None, None]])
+    return positions
+
+
+@pytest.fixture()
+def example_xy_data_negative() -> np.ndarray:
+    positions = np.array([[-1, -2, -3, -4], [-5, -6, -7, -8]])
+    return positions
+
+
+@pytest.fixture()
+def example_xy_data_float() -> np.ndarray:
+    positions = np.array(
+        [
+            [1.0, 2.3333, 0.00000000000000001, 99999999999999999],
+            [2.32843476297480273847, 6.0, 7.5, 8],
+        ]
+    )
+    return positions
+
+
+@pytest.fixture()
+def example_xy_data_string() -> np.ndarray:
+    positions = np.array([["1", "2", "3", "4"], ["5", "6", "7", "8"]])
+    return positions
